requirements.txt
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Automat==22.10.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.12.1
bleach==6.0.0
blinker==1.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
constantly==23.10.4
cryptography==42.0.1
cssselect==1.2.0
customtkinter==5.2.2
darkdetect==0.8.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==4.1.13
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.5.0
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.13.1
Flask==3.0.1
Flask-Cors==4.0.0
fqdn==1.5.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==7.0.1
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.14.0
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
lxml==5.1.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.1
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.2
notebook_shim==0.2.3
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
Protego==0.3.0
psutil==5.9.5
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==8.2.2
Pygments==2.16.1
pymongo==4.6.1
pyOpenSSL==24.0.0
pyppeteer==1.0.2
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytube==15.0.0
pytz==2023.3.post1
pywatchman==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
queuelib==1.6.2
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
Scrapy==2.11.0
Send2Trash==1.8.2
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.2.4
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tk==0.1.0
tldextract==5.1.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
Twisted==22.10.0
typing_extensions==4.9.0
uri-template==1.3.0
urllib3==1.26.18
virtualenv==20.25.0
w3lib==2.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
websockets==10.4
Werkzeug==3.0.1
zipp==3.17.0
zope.interface==6.1
